requirements.txt
# Core
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching
django-redis==5.4.0

# Auth & Security
django-cors-headers==4.3.1

# Filtering & Search
django-filter==24.2

# Static files
whitenoise==6.6.0

# HTTP (AI provider calls)
requests==2.31.0

# Server
gunicorn==21.2.0

# Dev / test only
pytest==8.1.1
pytest-django==4.8.0
factory-boy==3.3.0
